feat: revamp harbor cli
