+++ Makefile
@@ -1,0 +1,3 @@
+make: 
+	gofmt -l -s -w .
+	go build -o harbor cmd/harbor/main.go
