+++ cmd/harbor/root/project/create_project.go
@@ -3,10 +3,10 @@ package project
 import (
 	"context"
 
-	"github.com/akshatdalton/harbor-cli/cmd/constants"
-	"github.com/akshatdalton/harbor-cli/cmd/utils"
 	"github.com/goharbor/go-client/pkg/sdk/v2.0/client/project"
 	"github.com/goharbor/go-client/pkg/sdk/v2.0/models"
+	"github.com/goharbor/harbor-cli/pkg/constants"
+	"github.com/goharbor/harbor-cli/pkg/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -17,8 +17,8 @@ type createProjectOptions struct {
 	storageLimit int64
 }
 
-// NewCreateProjectCommand creates a new `harbor create project` command
-func NewCreateProjectCommand() *cobra.Command {
+// CreateProjectCommand creates a new `harbor create project` command
+func CreateProjectCommand() *cobra.Command {
 	var opts createProjectOptions
 
 	cmd := &cobra.Command{
