+++ cmd/harbor/root/project/get_project.go
@@ -3,23 +3,23 @@ package project
 import (
 	"context"
 
-	"github.com/akshatdalton/harbor-cli/cmd/constants"
-	"github.com/akshatdalton/harbor-cli/cmd/utils"
 	"github.com/goharbor/go-client/pkg/sdk/v2.0/client/project"
+	"github.com/goharbor/harbor-cli/pkg/constants"
+	"github.com/goharbor/harbor-cli/pkg/utils"
 	"github.com/spf13/cobra"
 )
 
-type deleteProjectOptions struct {
+type getProjectOptions struct {
 	projectNameOrID string
 }
 
-// NewDeleteProjectCommand creates a new `harbor delete project` command
-func NewDeleteProjectCommand() *cobra.Command {
-	var opts deleteProjectOptions
+// GetProjectCommand creates a new `harbor get project` command
+func GetProjectCommand() *cobra.Command {
+	var opts getProjectOptions
 
 	cmd := &cobra.Command{
 		Use:   "project [NAME|ID]",
-		Short: "delete project by name or id",
+		Short: "get project by name or id",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			opts.projectNameOrID = args[0]
@@ -27,17 +27,17 @@ func NewDeleteProjectCommand() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			return runDeleteProject(opts, credentialName)
+			return runGetProject(opts, credentialName)
 		},
 	}
 
 	return cmd
 }
 
-func runDeleteProject(opts deleteProjectOptions, credentialName string) error {
+func runGetProject(opts getProjectOptions, credentialName string) error {
 	client := utils.GetClientByCredentialName(credentialName)
 	ctx := context.Background()
-	response, err := client.Project.DeleteProject(ctx, &project.DeleteProjectParams{ProjectNameOrID: opts.projectNameOrID})
+	response, err := client.Project.GetProject(ctx, &project.GetProjectParams{ProjectNameOrID: opts.projectNameOrID})
 
 	if err != nil {
 		return err
