+++ cmd/harbor/root/registry/list_registry.go
@@ -3,9 +3,9 @@ package registry
 import (
 	"context"
 
-	"github.com/akshatdalton/harbor-cli/cmd/constants"
-	"github.com/akshatdalton/harbor-cli/cmd/utils"
 	"github.com/goharbor/go-client/pkg/sdk/v2.0/client/registry"
+	"github.com/goharbor/harbor-cli/pkg/constants"
+	"github.com/goharbor/harbor-cli/pkg/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -17,7 +17,7 @@ type listRegistryOptions struct {
 }
 
 // NewListRegistryCommand creates a new `harbor list registry` command
-func NewListRegistryCommand() *cobra.Command {
+func ListRegistryCommand() *cobra.Command {
 	var opts listRegistryOptions
 
 	cmd := &cobra.Command{
