+++ doc/doc.md
@@ -1,0 +1,1 @@
+We can create a worker function to automatically generate documents for users whenever a new release is available. This function will streamline the process of document generation and ensure that users have up-to-date information. By automating this task, we can improve efficiency and provide clear and concise documentation in a user-friendly manner. //@amands
