+++ test/e2e/test.md
@@ -1,0 +1,1 @@
+e2e tests for cmd can slide in here
